Moved matchandreplace and updated passes to differentiate between types of passes better
