+++ include/coreir-pass/common.h
@@ -1,7 +1,6 @@
 #ifndef PASSES_COMMON_H_
 #define PASSES_COMMON_H_
 
-//TODO Should probably move common passes elsewhere
-#include "../../src/lib/passes/matchandreplace.hpp"
+#include "matchandreplace.h"
 
 #endif //PASSES_HPP_
